missing Variant property?
