+++ BESM3CAData/Listings/DataListings/IVariantDataListing.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace BESM3CAData.Listings
+{
+    public interface IVariantDataListing
+    {
+        bool RequiresVariant { get; }
+        List<VariantListing> Variants { get; set; }
+    }
+}
