+++ BESM3CAData/Listings/DataListings/MultiGenreDataListing.cs
@@ -4,12 +4,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BESM3CAData.Model;
+using BESM3CAData.Control;
 
 namespace BESM3CAData.Listings
 {
     public class MultiGenreDataListing : LevelableDataListing
     {
         public List<int> GenrePoints { get; private set; }
+
+        public override DataNode CreateNode(string notes, DataController controller, int level = 1, int pointAdj = 0)
+        {
+            return new MultiGenreDataNode(this, notes, controller, level, pointAdj);
+        }
 
         public override bool MultiGenre
         {
